+++ server/controllers/authController.js
@@ -21,11 +21,15 @@ exports.register = async (req, res) => {
 
         user = new User({ username, email, password });
 <<<<<<< HEAD
+<<<<<<< HEAD
         user.password = await bcrypt.hash(password, process.env.PASSWORD_SALT);
 =======
         const salt = await bcrypt.genSalt(10);
         user.password = await bcrypt.hash(password, salt);
 >>>>>>> 3350c9c (server added)
+=======
+        user.password = await bcrypt.hash(password, process.env.PASSWORD_SALT);
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
         await user.save();
 
         const payload = { user: { id: user.id, username: user.username } };
@@ -43,7 +47,7 @@ exports.login = async (req, res) => {
     try {
         const user = await User.findOne({ email });
         if (!user) return res.status(400).json({ message: 'User not existed. Register Now!' });
-
+        
         const isMatch = await bcrypt.compare(password, user.password);
         if (!isMatch) return res.status(400).json({ message: 'Invalid Password, Try again!' });