+++ server/server.js
@@ -9,6 +9,7 @@ const connectDB = require('./config/db');
 const cors = require('cors');
 const auth = require('./routes/authRoutes');
 <<<<<<< HEAD
+<<<<<<< HEAD
 const User = require('./routes/userRoute');
 const AdminRoutes = require("./routes/adminRoute");
 const places = require('./routes/placesRoutes');
@@ -16,6 +17,9 @@ const sendPasswordReset = require("./controllers/passwordResetController");
 const { CALLBACK_URL } = require('./utils/general');
 =======
 >>>>>>> 3350c9c (server added)
+=======
+const sendPasswordReset = require("./controllers/passwordResetController");
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
 
 const app = express();
 
@@ -75,6 +79,29 @@ app.get('/api/visitors', (req, res) => {
 const PORT = process.env.PORT || 8080;
 app.listen(PORT, () => console.log(`Server started on port ${PORT}`));
 =======
+
+// Send email for forgetting password
+app.post('/api/passwordResetEmail', async (req, res) => {
+    const { email } = req.body;
+    try {
+        await sendPasswordReset.sendPasswordReset(email);
+        res.status(200).send({email});
+    } catch (error) {
+        res.status(500).send({ message: error.message });
+    }
+  });
+
+// Send reset password
+app.post('/api/resetPassword', async (req, res) => {
+    const { id } = req.body;
+    const { password } = req.body;
+    try {
+        await sendPasswordReset.resetPassword(id, password);
+        res.status(200).send({id});
+    } catch (error) {
+        res.status(500).send(error.message);
+    }
+  });
 
 const PORT = process.env.PORT || 8080;
 app.listen(PORT, () => console.log(`Server started on port ${PORT}`));