+++ src/components/ForgotPasswordModal.js
@@ -1,5 +1,9 @@
 import { useForm } from 'react-hook-form';
+<<<<<<< HEAD
 import { Modal, Button, Form, Alert } from 'react-bootstrap';
+=======
+import { Modal, Button, Form, Alert, ModalBody } from 'react-bootstrap';
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
 import { useState } from 'react';
 import { useRouter } from 'next/router';
 import styles from '../styles/login.module.css'; // Import custom CSS for styling
@@ -19,6 +23,7 @@ export default function ForgotPasswordModal({ handleClose, show }) {
     const onSubmit = async (email, e) => {
         e.preventDefault();
         console.log('Form submitted'); // Debugging line
+<<<<<<< HEAD
         try {
             const response = await sendPasswordResetEmail(email);
             if (response == 200) {
@@ -28,6 +33,17 @@ export default function ForgotPasswordModal({ handleClose, show }) {
             console.error('Error during sending email:', error); // Debugging line
             setWarning(error.message);
         }
+=======
+         try {
+             const response = await sendPasswordResetEmail(email);
+             if (response == 200) {
+                router.push("/reset-password-feedback");
+             }
+         } catch (error) {
+             console.error('Error during sending email:', error); // Debugging line
+             setWarning(error.message);
+         }
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
     };
 
     const handleCloseModal = () => {
@@ -36,7 +52,11 @@ export default function ForgotPasswordModal({ handleClose, show }) {
     };
 
     return (
+<<<<<<< HEAD
         <Modal dialogClassName={styles['modal-dialog']} onHide={handleCloseModal} show={show} centered scrollable>
+=======
+        <Modal dialogClassName={styles['modal-dialog']} onHide={handleCloseModal} show={show}centered scrollable>
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
             <Modal.Header closeButton>
                 <Modal.Title>Forgot Your Password ?</Modal.Title>
             </Modal.Header>
@@ -47,18 +67,33 @@ export default function ForgotPasswordModal({ handleClose, show }) {
 
                 <div className="px-5">
                     <Form id="forgot-password-form" onSubmit={handleSubmit(onSubmit)} >
+<<<<<<< HEAD
                         <Form.Group className="mt-4 mb-4" controlId="formEmail">
                             <Form.Label className={styles['custom-label']}> Email Address</Form.Label>
                             <Form.Control className={styles["custom-input"]}
+=======
+                        <Form.Group className="mt-4 mb-4" controlId="formBasicEmail">
+                            <Form.Label> Email Address</Form.Label>
+                            <Form.Control
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
                                 type="email"
                                 placeholder="Enter your email address"
                                 {...register('email', { required: 'Email is required' })}
                                 isInvalid={touchedFields.email && errors.email}
+<<<<<<< HEAD
                             />
                             {errors.email && <Form.Text className="text-danger">{errors.email.message}</Form.Text>}
                         </Form.Group>
                         <div className="mb-4">
                             <Button variant="success" className={styles['custom-submit-button']} type="submit" onClick={() => trigger()}>Submit</Button>
+=======
+                                style={{ padding: '15px' }}
+                            />
+                            {errors.email && <Form.Text className="text-danger">{errors.email.message}</Form.Text>}
+                        </Form.Group>
+                        <div className="d-grid gap-2">
+                            <Button variant="success" style={{ padding: '18px', marginTop: '5px', marginBottom: '20px' }} type="submit" onClick={() => trigger()}>Submit</Button>
+>>>>>>> 89ea291 (Forgot Password logic and UI implemented using nodemailer (#28))
                         </div>
                     </Form>
                 </div>