+++ src/services/index.js
@@ -104,7 +104,7 @@ export async function setPassword(id, password) {
 }
 
 export async function fetchPlaces(location, travelStyle) {
-    const res = await my_fetch(`${process.env.NEXT_PUBLIC_API_URL}/places?location=${encodeURIComponent(location)}&travelStyle=${travelStyle}`);
+    const res = await my_fetch(`${API_BASE_URL}/places?location=${encodeURIComponent(location)}&travelStyle=${travelStyle}`);
     const data = await res.json();
 
     if (res.status === 200) {
@@ -115,7 +115,7 @@ export async function fetchPlaces(location, travelStyle) {
 }
 
 export async function fetchPlaceDetails(place_id) {
-    const res = await my_fetch(`${process.env.NEXT_PUBLIC_API_URL}/places/details?place_id=${place_id}`);
+    const res = await my_fetch(`${API_BASE_URL}/places/details?place_id=${place_id}`);
     const data = await res.json();
 
     if (res.status === 200) {
@@ -124,7 +124,6 @@ export async function fetchPlaceDetails(place_id) {
         throw new Error(data.error);
     }
 }
-
 export async function my_fetch(url, args) {
     const _args = {
         ...args,